requirements.txt
redis
elasticsearch>=8.0.0,<9.0.0
pandas
requests
numpy
python-dotenv
